+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureCreateInfoNV.gen.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
-using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -41,7 +40,7 @@ public unsafe partial struct AccelerationStructureCreateInfoNV
             "VK_VERSION_1_1",
         ]
     )]
-    public ulong CompactedSize;
+    public nuint CompactedSize;
 
     [SupportedApiProfile(
         "vulkan",
