+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureGeometryLinearSweptSpheresDataNV.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -45,7 +46,7 @@ public unsafe partial struct AccelerationStructureGeometryLinearSweptSpheresData
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong VertexStride;
+    public nuint VertexStride;
 
     [SupportedApiProfile(
         "vulkan",
@@ -67,7 +68,7 @@ public unsafe partial struct AccelerationStructureGeometryLinearSweptSpheresData
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong RadiusStride;
+    public nuint RadiusStride;
 
     [SupportedApiProfile(
         "vulkan",
@@ -89,7 +90,7 @@ public unsafe partial struct AccelerationStructureGeometryLinearSweptSpheresData
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong IndexStride;
+    public nuint IndexStride;
 
     [SupportedApiProfile(
         "vulkan",
