+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureGeometrySpheresDataNV.gen.cs
@@ -46,7 +46,7 @@ public unsafe partial struct AccelerationStructureGeometrySpheresDataNV
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong VertexStride;
+    public nuint VertexStride;
 
     [SupportedApiProfile(
         "vulkan",
@@ -68,7 +68,7 @@ public unsafe partial struct AccelerationStructureGeometrySpheresDataNV
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong RadiusStride;
+    public nuint RadiusStride;
 
     [SupportedApiProfile(
         "vulkan",
@@ -90,5 +90,5 @@ public unsafe partial struct AccelerationStructureGeometrySpheresDataNV
         ["VK_NV_ray_tracing_linear_swept_spheres"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong IndexStride;
+    public nuint IndexStride;
 }
