+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureMatrixMotionInstanceNV.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -100,5 +101,5 @@ public partial struct AccelerationStructureMatrixMotionInstanceNV
         ["VK_NV_ray_tracing_motion_blur"],
         ImpliesSets = ["VK_KHR_ray_tracing_pipeline"]
     )]
-    public ulong AccelerationStructureReference;
+    public nuint AccelerationStructureReference;
 }
