+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureTrianglesOpacityMicromapEXT.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -59,7 +60,7 @@ public unsafe partial struct AccelerationStructureTrianglesOpacityMicromapEXT
             "VK_KHR_acceleration_structure+VK_VERSION_1_3",
         ]
     )]
-    public ulong IndexStride;
+    public nuint IndexStride;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile(
