+++ sources/Vulkan/Vulkan/Vulkan/AmigoProfilingSubmitInfoSEC.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -40,7 +41,7 @@ public unsafe partial struct AmigoProfilingSubmitInfoSEC
             "VK_SEC_amigo_profiling+VK_VERSION_1_1",
         ]
     )]
-    public ulong FirstDrawTimestamp;
+    public nuint FirstDrawTimestamp;
 
     [NativeTypeName("uint64_t")]
     [SupportedApiProfile(
@@ -51,5 +52,5 @@ public unsafe partial struct AmigoProfilingSubmitInfoSEC
             "VK_SEC_amigo_profiling+VK_VERSION_1_1",
         ]
     )]
-    public ulong SwapBufferTimestamp;
+    public nuint SwapBufferTimestamp;
 }
