+++ sources/Vulkan/Vulkan/Vulkan/AntiLagPresentationInfoAMD.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -21,5 +22,5 @@ public unsafe partial struct AntiLagPresentationInfoAMD
 
     [NativeTypeName("uint64_t")]
     [SupportedApiProfile("vulkan", ["VK_AMD_anti_lag"])]
-    public ulong FrameIndex;
+    public nuint FrameIndex;
 }
