+++ sources/Vulkan/Vulkan/Vulkan/BindImageMemoryInfo.gen.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Silk.NET.Vulkan;
 
@@ -46,5 +48,5 @@ public unsafe partial struct BindImageMemoryInfo
         ["VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.1"
     )]
-    public ulong MemoryOffset;
+    public nuint MemoryOffset;
 }
