+++ sources/Vulkan/Vulkan/Vulkan/BindPipelineIndirectCommandNV.gen.cs
@@ -17,5 +17,5 @@ public partial struct BindPipelineIndirectCommandNV
         ["VK_NV_device_generated_commands_compute"],
         ImpliesSets = ["VK_NV_device_generated_commands"]
     )]
-    public ulong PipelineAddress;
+    public nuint PipelineAddress;
 }
