+++ sources/Vulkan/Vulkan/Vulkan/BufferCreateInfo.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -39,7 +40,7 @@ public unsafe partial struct BufferCreateInfo
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public ulong Size;
+    public nuint Size;
 
     [NativeTypeName("VkBufferUsageFlags")]
     [SupportedApiProfile(
