+++ sources/Vulkan/Vulkan/Vulkan/BufferUsageFlags2CreateInfo.gen.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
@@ -19,5 +20,5 @@ public unsafe partial struct BufferUsageFlags2CreateInfo
 
     [NativeTypeName("VkBufferUsageFlags2")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_4"], MinVersion = "1.4")]
-    public ulong Usage;
+    public nuint Usage;
 }
