+++ sources/Vulkan/Vulkan/Vulkan/CheckpointData2NV.gen.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // Ported from the Vulkan headers and corresponding dependencies.
 // Original source is Copyright 2015-2023 The Khronos Group Inc. Licensed under the MIT license.
+using System;
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace Silk.NET.Vulkan;
 
@@ -47,7 +49,7 @@ public unsafe partial struct CheckpointData2NV
             "VK_NV_device_diagnostic_checkpoints+VK_VERSION_1_1",
         ]
     )]
-    public ulong Stage;
+    public nuint Stage;
 
     [SupportedApiProfile(
         "vulkan",
