+++ sources/Vulkan/Vulkan/Vulkan/ClusterAccelerationStructureBuildTriangleClusterInfoNV.gen.cs
@@ -149,7 +149,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public ulong IndexBuffer;
+    public nuint IndexBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -157,7 +157,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public ulong VertexBuffer;
+    public nuint VertexBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -165,7 +165,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public ulong GeometryIndexAndFlagsBuffer;
+    public nuint GeometryIndexAndFlagsBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -173,7 +173,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public ulong OpacityMicromapArray;
+    public nuint OpacityMicromapArray;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -181,5 +181,5 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public ulong OpacityMicromapIndexBuffer;
+    public nuint OpacityMicromapIndexBuffer;
 }
