+++ sources/Vulkan/Vulkan/Vulkan/DebugReportCallbackCreateInfoEXT.gen.cs
@@ -27,7 +27,7 @@ public unsafe partial struct DebugReportCallbackCreateInfoEXT
     public delegate* unmanaged<
         uint,
         DebugReportObjectTypeEXT,
-        ulong,
+        nuint,
         nuint,
         int,
         sbyte*,
